pyproject.toml
[project]
name = "vrp"
version = "0.1.0"
description = ""
authors = [
    {name = "Peter",email = "ptbdnr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]

[tool.poetry]
packages = [{include = "poetry_tsp", from = "src"}]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
